+++ .github/workflows/pull-request.yml
@@ -17,13 +17,6 @@ jobs:
           packageUser: ${{ github.actor }}
           packagePAT: ${{ secrets.GITHUB_TOKEN }}
         run: ./gradlew build
-      - name: Archive Error Log
-        uses: actions/upload-artifact@v2
-        if: failure()
-        with:
-          name: Ballerina Internal Log
-          path: random-ballerina/ballerina-internal.log
-          if-no-files-found: ignore
       - name: Generate CodeCov Report
         if: github.event_name == 'pull_request'
         uses: codecov/codecov-action@v1
