+++ ballerina/natives.bal
@@ -37,5 +37,5 @@ public isolated function createDecimal() returns float = @java:Method {
 # + return - Selected random value or else a `random:Error` if the start range is greater than the end range
 public isolated function createIntInRange(int startRange, int endRange) returns int|Error = @java:Method {
     name: "randomInRange",
-    'class: "org.ballerinalang.stdlib.random.nativeimpl.ExternMethods"
+    'class: "io.ballerina.stdlib.random.nativeimpl.ExternMethods"
 } external;
