+++ gradle.properties
@@ -1,7 +1,7 @@
 org.gradle.caching=true
-group=org.ballerinalang
+group=io.ballerina.stdlib
 version=0.10.0-beta.3-SNAPSHOT
 
-ballerinaGradlePluginVersion=0.9.1
+ballerinaGradlePluginVersion=0.10.0
 
 ballerinaLangVersion=2.0.0-beta.2.1
